+++ packages/hooks_riverpod/lib/misc.dart
@@ -22,4 +22,5 @@ export 'src/internals.dart'
         ProviderTransformerContext,
         ProviderTransformer,
         SyncProviderTransformerMixin,
-        AsyncProviderTransformerMixin;
+        AsyncProviderTransformerMixin,
+        AsyncProviderListenable;