+++ packages/riverpod/lib/misc.dart
@@ -21,4 +21,5 @@ export 'src/internals.dart'
         ProviderTransformerContext,
         ProviderTransformer,
         SyncProviderTransformerMixin,
-        AsyncProviderTransformerMixin;
+        AsyncProviderTransformerMixin,
+        AsyncProviderListenable;